+++ cmd/livesim2/app/asset_test.go
@@ -28,6 +28,7 @@ type wantedRepData struct {
 	mpdTimescale   int // SegmentTemplate timescale
 	mediaTimescale int
 	duration       int
+	editListOffset int64 // Offset of the elst box in the init segment
 }
 
 func TestLoadAsset(t *testing.T) {
@@ -39,6 +40,31 @@ func TestLoadAsset(t *testing.T) {
 		ad           wantedAssetData
 		rds          map[string]wantedRepData
 	}{
+		{
+			desc:      "CTA-Wave AAC with editlist",
+			assetPath: "assets/WAVE/av",
+			ad: wantedAssetData{
+				nrReps:         2,
+				loopDurationMS: 8000,
+			},
+			rds: map[string]wantedRepData{
+				"video25fps": {
+					nrSegs:         4,
+					initURI:        "video25fps/init.mp4",
+					mpdTimescale:   12_800,
+					mediaTimescale: 12_800,
+					duration:       102_400,
+				},
+				"aac": {
+					nrSegs:         5, // To get longer duration than video
+					initURI:        "aac/init.mp4",
+					mpdTimescale:   48_000,
+					mediaTimescale: 48_000,
+					duration:       476160, // 9.92s which is fine since longer than video
+					editListOffset: 2048,
+				},
+			},
+		},
 		{
 			desc:         "testpic_2s",
 			assetPath:    "assets/testpic_2s",
@@ -116,17 +142,18 @@ func TestLoadAsset(t *testing.T) {
 				asset, ok := am.findAsset(tc.assetPath)
 				require.True(t, ok)
 				require.NotNil(t, asset)
-				require.Equal(t, tc.ad.nrReps, len(asset.Reps))
+				require.Equal(t, tc.ad.nrReps, len(asset.Reps), "nr reps in asset %s", asset.AssetPath)
 				require.Equal(t, tc.ad.loopDurationMS, asset.LoopDurMS)
 				for repID, wrd := range tc.rds {
 					rep, ok := asset.Reps[repID]
-					require.True(t, ok)
+					require.True(t, ok, "repID %s not found in asset %s", repID, asset.AssetPath)
 					require.NotNil(t, rep)
-					require.Equal(t, wrd.nrSegs, len(rep.Segments))
+					require.Equal(t, wrd.nrSegs, len(rep.Segments), "repID %s in asset %s", repID, asset.AssetPath)
 					require.Equal(t, wrd.initURI, rep.InitURI)
-					require.Equal(t, wrd.mpdTimescale, rep.MpdTimescale)
-					require.Equal(t, wrd.mediaTimescale, rep.MediaTimescale)
+					require.Equal(t, wrd.mpdTimescale, rep.MpdTimescale, "repID %s in asset %s", repID, asset.AssetPath)
+					require.Equal(t, wrd.mediaTimescale, rep.MediaTimescale, "repID %s in asset %s", repID, asset.AssetPath)
 					require.Equal(t, wrd.duration, rep.duration())
+					require.Equal(t, wrd.editListOffset, rep.EditListOffset)
 				}
 			}
 		})